Refactor schema core into storage and validator modules
