+++ fold_node/src/schema/mod.rs
@@ -23,11 +23,13 @@
 
 // Internal modules
 pub(crate) mod core;
+pub(crate) mod storage;
 pub mod types;
 
 
 // Public re-exports
 pub use core::SchemaCore;
+pub use storage::SchemaStorage;
 pub use types::{errors::SchemaError, schema::Schema, Transform};
 pub mod validator;
 pub use validator::SchemaValidator;
